Improved the entry api + refactored sorting
This first goal of the commit was to make it so that `&str` could be
used to search for `String` keys. All of my testing had been against
numeric keys, and I had wrongly assumed that `String` implemented
`PartialOrd<str>` (<https://github.com/rust-lang/rust/issues/82990>).

The new `Sort` trait works around this in better ways than I had
expected -- not only can it provide automatic implementations for all
Ord/PartialOrd implementers, it also can have specialized
implementations for String/Vec. This means that end users will only have
to implement Ord/PartialOrd for their key types, and almost no one
should need to manually implement this trait.

The next goal was to change the entry API into a form that accepts owned or
borrowed representations of the `Key` type and delays the conversion to
the owned form until after the key isn't found in the collection. The
net effect is that passing a str into entry() will use the borrowed
representation to search for the key, and only convert to a String if
the key isn't found. And, when the key is already in its owned
representation, no extra clone is made.